requirements.txt
tweepy==4.14.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
nltk==3.8.1
textblob==0.17.1
redis==4.6.0
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
asyncio-mqtt==0.16.1
schedule==1.2.0
loguru==0.7.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
slack-sdk==3.22.0 